# Optimize small read on ORC(small row group index/stripe/file) (#3962)