--- a/be/src/exec/vectorized/hdfs_scanner_orc.cpp
+++ b/be/src/exec/vectorized/hdfs_scanner_orc.cpp
@@ -15,7 +15,7 @@ class ORCHdfsFileStream : public orc::InputStream {
 public:
     // |file| must outlive ORCHdfsFileStream
     ORCHdfsFileStream(RandomAccessFile* file, uint64_t length, HdfsScanStats* stats)
-            : _file(std::move(file)), _length(length), _stats(stats) {}
+            : _file(std::move(file)), _length(length), _stats(stats), _cache_buffer(0), _cache_offset(0) {}
 
     ~ORCHdfsFileStream() override = default;
 
@@ -40,7 +40,40 @@ public:
 
     uint64_t getNaturalReadSizeAfterSeek() const override { return 256 * 1024; }
 
+    void prepareCache(orc::InputStream::PrepareCacheScope scope, uint64_t offset, uint64_t length) override {
+        const size_t cache_max_size = config::orc_file_cache_max_size;
+        if (length > cache_max_size) return;
+        if (canUseCacheBuffer(offset, length)) return;
+
+        // If this stripe is small, probably other stripes are also small
+        // we combine those reads into one, and try to read several stripes in one shot.
+        if (scope == orc::InputStream::PrepareCacheScope::READ_FULL_STRIPE) {
+            length = std::min(_length - offset, cache_max_size);
+        }
+
+        _cache_buffer.resize(length);
+        _cache_offset = offset;
+        doRead(_cache_buffer.data(), length, offset);
+    }
+
+    inline bool canUseCacheBuffer(uint64_t offset, uint64_t length) {
+        if ((_cache_buffer.size() != 0) && (offset >= _cache_offset) &&
+            ((offset + length) <= (_cache_offset + _cache_buffer.size()))) {
+            return true;
+        }
+        return false;
+    }
+
     void read(void* buf, uint64_t length, uint64_t offset) override {
+        if (canUseCacheBuffer(offset, length)) {
+            size_t idx = offset - _cache_offset;
+            memcpy(buf, _cache_buffer.data() + idx, length);
+        } else {
+            doRead(buf, length, offset);
+        }
+    }
+
+    void doRead(void* buf, uint64_t length, uint64_t offset) {
         SCOPED_RAW_TIMER(&_stats->io_ns);
         _stats->io_count += 1;
         if (buf == nullptr) {
@@ -61,6 +94,8 @@ private:
     RandomAccessFile* _file;
     uint64_t _length;
     HdfsScanStats* _stats;
+    std::vector<char> _cache_buffer;
+    uint64_t _cache_offset;
 };
 
 class OrcRowReaderFilter : public orc::RowReaderFilter {