--- a/be/src/formats/orc/apache-orc/c++/include/orc/OrcFile.hh
+++ b/be/src/formats/orc/apache-orc/c++/include/orc/OrcFile.hh
@@ -39,6 +39,12 @@ namespace orc {
    */
 class InputStream {
 public:
+    enum class PrepareCacheScope {
+        READ_FULL_FILE,
+        READ_FULL_STRIPE,
+        READ_ROW_GROUP_INDEX,
+    };
+
     virtual ~InputStream();
 
     /**
@@ -71,6 +77,8 @@ public:
      * Get the name of the stream for error messages.
      */
     virtual const std::string& getName() const = 0;
+
+    virtual void prepareCache(PrepareCacheScope scope, uint64_t offset, uint64_t length);
 };
 
 /**