This PR is to optimization read small orc file and small stripe. And what's more, we can put previous optimization on row group index into this framework.

Handle Small File
=======
I set up a test environment of a lot of small hdfs files. All of them are small files under 2MB.

```
-rw-r-----   3 hadoop hadoop    1332877 2022-03-09 12:10 /user/zya/tpch_100g/lineitem_smallfile_zlib/012314_0
-rw-r-----   3 hadoop hadoop    1328418 2022-03-09 12:11 /user/zya/tpch_100g/lineitem_smallfile_zlib/012315_0
-rw-r-----   3 hadoop hadoop    1312973 2022-03-09 12:11 /user/zya/tpch_100g/lineitem_smallfile_zlib/012316_0
-rw-r-----   3 hadoop hadoop    1304832 2022-03-09 12:11 /user/zya/tpch_100g/lineitem_smallfile_zlib/012317_0
-rw-r-----   3 hadoop hadoop     878983 2022-03-09 12:10 /user/zya/tpch_100g/lineitem_smallfile_zlib/012318_0
```

Running query
> select count(*) from lineitem_smallfile;

You can see although IOBytes increased, but IOCounte decreased. And average bytes per IO is increase from (629.1MB / 257.963K ~= 2.69KB) to (16.72GB / 12.319K ~= 1.35MB)

Before optimization

```
         HDFS_SCAN_NODE (id=0):(Active: 8s861ms[8861568366ns], % non-child: 98.73%)
             - BytesRead: 0.00
             - BytesReadDataNodeCache: 0.00
             - BytesReadFromDisk: 629.10 MB
             - BytesReadLocal: 0.00
             - BytesReadRemote: 629.10 MB
             - BytesReadShortCircuit: 0.00
             - BytesTotalRead: 629.10 MB
             - IoCounter: 257.063K (257063)
             - IoTime: 6m10s
```

After optimization
```
          HDFS_SCAN_NODE (id=0):(Active: 4s322ms[4322893106ns], % non-child: 97.17%)
             - Table: lineitem_smallfile
             - Predicates:
             - PredicatesPartition:
             - BytesRead: 16.72 GB
             - ColumnConvertTime: 1s250ms
             - ColumnReadTime: 16s355ms
             - ExprFilterTime: 8.39ms
             - HdfsIO: 0ns
               - TotalBytesRead: 17.957812729B (17957812729)
               - TotalLocalBytesRead: 0
               - TotalShortCircuitBytesRead: 0
               - TotalZeroCopyBytesRead: 0
             - IoCounter: 12.319K (12319) // <=== exact file number
             - IoTime: 2m17s
```

Handle Stripe Size
=======
I set up a test environment of a single file but with small stripe size. You can see a stripe size is abot 700K, and there are 350 stripes in a single file.

> java -jar ~/installed/orc-tools-1.7.0-SNAPSHOT-uber.jar meta test_many_columns.orc

```
  Stripe 348:
  Stripe 349:
  Stripe 350:
Stripes:
  Stripe: offset: 3 data: 722325 rows: 2864 tail: 1861 index: 9095
  Stripe: offset: 733284 data: 763098 rows: 2864 tail: 1907 index: 9096
  Stripe: offset: 1507385 data: 776631 rows: 2864 tail: 1880 index: 9097
```

Normally we don't have such small stripe size. Usually stripe size is about 256MB. But there are some bugs in SparkSQL may use small stripe size.
- https://community.cloudera.com/t5/Support-Questions/Spark-ORC-Stripe-Size/td-p/189844
- https://issues.apache.org/jira/browse/HIVE-13232

Running query
> select max(c1) from test_many_columns;

You can see IOCounter is cut from 2451 down to 351.

Before optimization

```
          HDFS_SCAN_NODE (id=0):(Active: 2s771ms[2771719221ns], % non-child: 99.20%)
             - BytesRead: 0.00
             - BytesReadDataNodeCache: 0.00
             - BytesReadFromDisk: 1.92 MB
             - BytesReadLocal: 0.00
             - BytesReadRemote: 1.92 MB
             - BytesReadShortCircuit: 0.00
             - BytesTotalRead: 1.92 MB
             - ColumnConvertTime: 20.341ms
             - ColumnReadTime: 1s814ms
             - ExprFilterTime: 0ns
             - GroupChunkRead: 0ns
             - GroupDictDecode: 0ns
             - GroupDictFilter: 0ns
             - IoCounter: 2.451K (2451)
             - IoTime: 1s772ms
```

After optimization

```
          HDFS_SCAN_NODE (id=0):(Active: 697.474ms[697474842ns], % non-child: 98.24%)
             - Table: test_many_columns
             - Predicates:
             - PredicatesPartition:
             - BytesRead: 255.89 MB
             - ColumnConvertTime: 20.537ms
             - ColumnReadTime: 676.219ms
             - ExprFilterTime: 15.637us
             - HdfsIO: 0ns
               - TotalBytesRead: 268.315441M (268315441)
               - TotalLocalBytesRead: 0
               - TotalShortCircuitBytesRead: 0
               - TotalZeroCopyBytesRead: 0
             - IoCounter: 351 // <== exact stripe size + file footer read.
             - IoTime: 612.488ms
```

And if we combine reads of stripes, we can save more IOCounter

```
          HDFS_SCAN_NODE (id=0):(Active: 643.709ms[643709229ns], % non-child: 97.69%)
             - Table: test_many_columns
             - Predicates:
             - PredicatesPartition:
             - BytesRead: 349.31 MB
             - ColumnConvertTime: 22.263ms
             - ColumnReadTime: 623.502ms
             - ExprFilterTime: 11.235us
             - HdfsIO: 0ns
               - TotalBytesRead: 366.282951M (366282951)
               - TotalLocalBytesRead: 0
               - TotalShortCircuitBytesRead: 0
               - TotalZeroCopyBytesRead: 0
             - IoCounter: 176
             - IoTime: 564.540ms
 ```

## What type of PR is this：
- [ ] bug
- [ ] feature
- [ ] enhancement
- [ ] others

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->